Do not require mocks for ARRAY JOIN clause arguments
The `ARRAY JOIN` clause in Clickhouse is currently recognised as a table
that requires a mock by the library. This is not the case as this clause
is used for expanding arrays in the source table into separate rows.
This commit adds an additional clause to the `get_source_tables` helper
function to ignore joins of kind `ARRAY`.

Closes #48